requirements.txt
python-dotenv
pandas
numpy
scikit-learn
scipy
matplotlib
seaborn
setuptools
pymongo
certifi
pymongo[srv] == 4.7
dill
pyyaml
mlflow==2.22.1
dagshub
fastapi
uvicorn
python-multipart
jinja2

##-e .